server/requirements.txt
astroid==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
django-crispy-forms==1.7.0
django-filter==1.1.0
djangorestframework==3.7.1
httpie==0.9.9
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.3.2
Pygments==2.2.0
pylint==1.7.4
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
